# swagger/spec.json
{"openapi":"3.0.0","info":{"title":"CoupDeMain API","version":"1.0.0"},"paths":{"/address":{"get":{"tags":["Address"],"security":[{"bearerAuth":[]}],"responses":{"401":{"$ref":"#/components/responses/MissingJWT"},"404":{"description":"Opération échouée"},"500":{"description":"Erreur serveur"}}},"patch":{"tags":["Address"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/UpdateAddress"},"responses":{"200":{"description":"Adresse mise à jour"},"400":{"description":"Paramètres invalides"},"401":{"$ref":"#/components/responses/MissingJWT"},"404":{"description":"Client invalide"},"500":{"description":"Erreur serveur"}}},"post":{"tags":["Address"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/CreateAddress"},"responses":{"201":{"description":"Adresse mise à jour"},"400":{"description":"Adresse inexistante"},"401":{"$ref":"#/components/responses/MissingJWT"},"500":{"description":"Erreur serveur"}}},"delete":{"tags":["Address"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/DeleteClient"},"responses":{"204":{"description":"Adresse supprimée"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeManager"},"404":{"description":"L'adresse n'existe pas"},"500":{"description":"Erreur serveur"}}}},"/address/{id}":{"get":{"tags":["Address"],"security":[{"bearerAuth":[]}],"parameters":[{"name":"id","description":"ID de l'adresse","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Adresse trouvée"},"400":{"description":"Paramètre incorrect"},"401":{"$ref":"#/components/responses/MissingJWT"},"404":{"description":"Adresse non trouvée"},"500":{"description":"Erreur serveur"}}}},"/announce":{"get":{"tags":["Announce"],"security":[{"bearerAuth":[]}],"responses":{"401":{"$ref":"#/components/responses/MissingJWT"},"404":{"description":"Opération échouée"},"500":{"description":"Erreur serveur"}}},"patch":{"tags":["Announce"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/UpdateAnnounce"},"responses":{"200":{"description":"Annonce mise à jour"},"400":{"description":"Date non antérieure"},"401":{"$ref":"#/components/responses/MissingJWT"},"404":{"description":"Annonce invalide"},"500":{"description":"Erreur serveur"}}},"post":{"tags":["Announce"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/CreateAnnounce"},"responses":{"201":{"description":"Annonce créée"},"400":{"description":"Annonce inexistante"},"401":{"$ref":"#/components/responses/MissingJWT"},"500":{"description":"Erreur serveur"}}},"delete":{"tags":["Announce"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/DeleteAnnounce"},"responses":{"204":{"description":"Annonce supprimée"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeManager"},"404":{"description":"L'annonce n'existe pas"},"500":{"description":"Erreur serveur"}}}},"/announce/{id}":{"get":{"tags":["Announce"],"security":[{"bearerAuth":[]}],"parameters":[{"name":"id","description":"ID d'une annonce","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Annonce trouvée"},"400":{"description":"Paramètre incorrect"},"401":{"$ref":"#/components/responses/MissingJWT"},"404":{"description":"Annonce non trouvée"},"500":{"description":"Erreur serveur"}}}},"/announce/{id}/appointments":{"get":{"tags":["Announce"],"security":[{"bearerAuth":[]}],"parameters":[{"name":"id","description":"ID d'une annonce","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"400":{"description":"Paramètre incorrect"},"401":{"$ref":"#/components/responses/MissingJWT"},"404":{"description":"Annonce inexistante"},"500":{"description":"Erreur serveur"}}}},"/appointment":{"get":{"tags":["Appointment"],"security":[{"bearerAuth":[]}],"responses":{"401":{"$ref":"#/components/responses/MissingJWT"},"404":{"description":"Opération échouée"},"500":{"description":"Erreur serveur"}}},"post":{"tags":["Appointment"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/CreateAppointment"},"responses":{"201":{"description":"Rendez-vous créé"},"400":{"description":"Paramètres inccorects"},"401":{"$ref":"#/components/responses/MissingJWT"},"404":{"description":"Erreur de création"},"500":{"description":"Erreur serveur"}}},"patch":{"tags":["Appointment"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/UpdateAppointment"},"responses":{"204":{"description":"Rendez-vous mis à jour"},"400":{"description":"Status invalide"},"401":{"$ref":"#/components/responses/MissingJWT"},"404":{"description":"Rendez-vous inexistant"},"500":{"description":"Erreur serveur"}}},"delete":{"tags":["Appointment"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/DeleteAppointment"},"responses":{"204":{"$ref":"#/components/responses/AppointmentDeleted"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeManager"},"404":{"description":"Le rendez-vous n'existe pas"},"500":{"description":"Erreur serveur"}}}},"/appointments/{id}":{"get":{"tags":["Appointment"],"security":[{"bearerAuth":[]}],"parameters":[{"name":"id","description":"ID d'un rendez-vous","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Rendez-vous trouvé"},"401":{"$ref":"#/components/responses/MissingJWT"},"404":{"description":"Rendez-vous non trouvé"},"500":{"description":"Erreur serveur"}}}},"/category":{"get":{"tags":["Category"],"security":[{"bearerAuth":[]}],"responses":{"401":{"$ref":"#/components/responses/MissingJWT"},"404":{"description":"Opération échouée"},"500":{"description":"Erreur serveur"}}}},"/category/{id}":{"get":{"tags":["Category"],"security":[{"bearerAuth":[]}],"parameters":[{"name":"nom","description":"Nom d'une catégorie","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Catégorie trouvée"},"401":{"$ref":"#/components/responses/MissingJWT"},"404":{"description":"Catégorie non trouvée"},"500":{"description":"Erreur serveur"}}}},"/entity":{"get":{"tags":["Client"],"security":[{"bearerAuth":[]}],"responses":{"401":{"$ref":"#/components/responses/MissingJWT"},"404":{"description":"Opération échouée"},"500":{"description":"Erreur serveur"}}},"post":{"tags":["Client"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/CreateClient"},"responses":{"201":{"$ref":"#/components/responses/ClientCreated"},"400":{"description":"L'email est déjà utilisé ou le format du mot de passe est incorrect"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeManager"},"500":{"description":"Erreur serveur"}}},"patch":{"tags":["Client"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/ClientAUpdate"},"responses":{"204":{"$ref":"#/components/responses/ClientUpdated"},"401":{"$ref":"#/components/responses/MissingJWT"},"500":{"description":"Erreur serveur"}}},"delete":{"tags":["Client"],"security":[{"bearerAuth":[]}],"requestBody":{"$ref":"#/components/requestBodies/DeleteClient"},"responses":{"204":{"$ref":"#/components/responses/ClientDeleted"},"401":{"$ref":"#/components/responses/MissingJWT"},"403":{"$ref":"#/components/responses/mustBeManager"},"404":{"description":"Le client n'existe pas"},"500":{"description":"Erreur serveur"}}}},"/entity/{id}":{"get":{"tags":["Client"],"security":[{"bearerAuth":[]}],"parameters":[{"name":"id","description":"ID d'un client","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"200":{"description":"Client trouvé"},"401":{"$ref":"#/components/responses/MissingJWT"},"404":{"description":"Client non trouvé"},"500":{"description":"Erreur serveur"}}}},"/entity/{id}/announces":{"get":{"summary":"Return announces of client","tags":["Client"],"security":[{"bearerAuth":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer"}}],"responses":{"400":{"description":"Paramètre incorrect"},"401":{"$ref":"#/components/responses/MissingJWT"},"404":{"description":"Client non trouvé"},"500":{"description":"Erreur serveur"}}}},"/login":{"post":{"tags":["Client"],"description":"Renvoie un JWT token permettant l'identification","requestBody":{"$ref":"#/components/requestBodies/ConnectClient"},"responses":{"200":{"description":"un token JWT","content":{"text/plain":{"schema":{"type":"string"}}}},"400":{"description":"Arguments manquants ou invalides"},"500":{"description":"Erreur serveur"}}}},"/register":{"post":{"tags":["Client"],"requestBody":{"$ref":"#/components/requestBodies/CreateClient"},"responses":{"201":{"$ref":"#/components/responses/ClientCreated"},"400":{"description":"L'email est déjà utilisé ou le format du mot de passe est incorrect"},"500":{"description":"Erreur serveur"}}}},"/role":{"get":{"tags":["Role"],"security":[{"bearerAuth":[]}],"responses":{"401":{"$ref":"#/components/responses/MissingJWT"},"404":{"description":"Opération échouée"},"500":{"description":"Erreur serveur"}}}}},"components":{"requestBodies":{"CreateAddress":{"content":{"application/json":{"schema":{"properties":{"userId":{"type":"integer"},"street":{"type":"string"},"city":{"type":"string"},"zipCode":{"type":"string"}}}}}},"UpdateAddress":{"content":{"application/json":{"schema":{"properties":{"id":{"type":"integer","description":"ID de l'adresse"},"street":{"type":"string"},"city":{"type":"string"},"zipCode":{"type":"string"}}}}}},"DeleteAddress":{"content":{"application/json":{"schema":{"properties":{"id":{"type":"integer"}}}}}},"UpdateAnnounce":{"content":{"application/json":{"schema":{"properties":{"id":{"type":"integer"},"title":{"type":"string"},"description":{"type":"string"},"date":{"type":"string","format":"date-time"},"duration":{"type":"string"},"price":{"type":"number","format":"float"},"categoryName":{"type":"string"},"entityId":{"type":"integer"},"statusId":{"type":"integer"}}}}}},"CreateAnnounce":{"content":{"application/json":{"schema":{"properties":{"title":{"type":"string"},"description":{"type":"string"},"date":{"type":"string","format":"date-time"},"duration":{"type":"integer"},"price":{"type":"number","format":"float"},"city":{"type":"string"},"categoryName":{"type":"string"},"entityId":{"type":"integer"}}}}}},"DeleteAnnounce":{"content":{"application/json":{"schema":{"properties":{"id":{"type":"integer"}}}}}},"CreateAppointment":{"content":{"application/json":{"schema":{"properties":{"announceId":{"type":"integer"},"entityId":{"type":"integer","description":"ID de l'auteur de la demande de rendez-vous"}}}}}},"UpdateAppointment":{"content":{"application/json":{"schema":{"properties":{"id":{"type":"integer","description":"ID du rendez-vous"},"statusId":{"type":"integer","description":"6= Openned, 7= Accepted, 8= Denied, 9= Canceled"}}}}}},"DeleteAppointment":{"content":{"application/json":{"schema":{"properties":{"id":{"type":"integer"}}}}}},"ConnectClient":{"description":"Identifiants de connexion","content":{"application/json":{"schema":{"properties":{"email":{"type":"string","format":"email"},"password":{"type":"string","format":"password"}}}}}},"CreateClient":{"content":{"application/json":{"schema":{"properties":{"email":{"type":"string","format":"email"},"firstname":{"type":"string"},"lastname":{"type":"string"},"password":{"type":"string","format":"password"}}}}}},"ClientAUpdate":{"content":{"application/json":{"schema":{"properties":{"id":{"type":"integer"},"email":{"type":"string","format":"email"},"firstname":{"type":"string"},"lastname":{"type":"string"},"password":{"type":"string","format":"password"},"description":{"type":"string"},"phoneNumber":{"type":"string"}}}}}},"DeleteClient":{"content":{"application/json":{"schema":{"properties":{"id":{"type":"integer"}}}}}}},"responses":{"AppointmentDeleted":{"description":"Rendez-vous supprimé"},"ClientCreated":{"description":"le client a été créé"},"ClientUpdated":{"description":"le client a été mis à jour"},"ClientDeleted":{"description":"le client a été supprimé"},"ErrorJWT":{"description":"le JWT n'est pas valide"},"MissingJWT":{"description":"le JWT n'est pas présent"},"mustBeManager":{"description":"L'action demandée ne peut être réalisée que par un manager"}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"tags":[]}